bot/requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
aiohttp==3.9.3
psycopg2-binary==2.9.9
pytz==2024.1
# --- Версии для ИИ, которые точно совместимы ---
groq==0.9.0
google-generativeai==0.7.1
# 🔥 КЛЮЧЕВОЕ ИСПРАВЛЕНИЕ: Жестко фиксируем версии httpx и httpcore
httpx==0.27.0
httpcore==1.0.5
